requirements.txt
tqdm 
torch
accelerate
peft
openpyxl
opencv-python
qwen-vl-utils
torchvision
jax
transformers==4.46.1
tokenizers==0.20.3
scikit-learn
Appium-Python-Client
pyyaml
pandas
gradio
spaces
omegaconf
matplotlib
einops 
transformers_stream_generator
tiktoken